seperated params and legacy params
